refactor: introduce new API response classes for enhanced response handling
- Added `ApiResponse<TData>` for detailed API responses with data.
- Added `PaginatedApiResponse<TData>` for paginated responses including pagination info.
- Introduced `SimpleApiResponse` for basic response structure.
- Created `ResponseWrapper` utility class for consistent API response formatting.
- Removed deprecated `Result` class to streamline response management.
